relayer: seal eras only on startup in a separate task
